AB_Testing_Project
This project is about analytical task from an international online store that launched an A/B test and then quit. We were only given the technical specifications and the test results. The test was named recommender_system_test and involved two groups - control group A and experimental group B, which introduced a new payment funnel. The test lasted about a year (December 2020 - January 2021), and the target audience of the test was 15% of new users from the EU region.

The purpose of the test was to check changes related to the introduction of an improved recommendation system, We expect to see a better conversion rate for users who sign up within 14 days into product page views (product_page), product card views (product_card), and purchases (purchase), it was expected that there would be at least a 10% increase in each stage of the funnel (product_page → product_card → purchase).

In order to answer all these questions,first I have to load the data, restore and clean it as necessary.check if there are missing values, duplicates. I might need to fill missing values with grouping data from more than one column to get more specefied calculations to move forward with the code.