SQL_Project
I've been given a database of one of the services competing in books apps market. It contains data on books, publishers, authors, customer ratings and reviews of books. My task is to analyze them and to generate a value proposition for a new product.

In order to do that firstly I need to:

-Study the tables by printing them, check if there are missing values and duplicates.

-Answering the questions bellow by making SQL queries for each task, and outputting the results.

-Describing my conclusions for each of the tasks.